+++ arch/x86_32/core/gdt.c
@@ -19,10 +19,11 @@ along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
 #include <stdint.h>
 
-#include "arch/x86_32/tss.h"
+#include "include/tss.h"
+
 #include "memory/vmm.h"
 
-#include "arch/x86_32/gdt.h"
+#include "include/gdt.h"
 
 // Defined in asm/gdt_flush.asm
 extern void gdt_flush(uint32_t gdt_ptr_addr);