+++ arch/x86_32/core/keyboard.c
@@ -20,7 +20,8 @@ along with this program.  If not, see <https://www.gnu.org/licenses/>.
 #include <stdbool.h>
 #include <stddef.h>
 
-#include "arch/stubs.h"
+#include "hal.h"
+
 #include "drivers/terminal.h"
 
 #include "drivers/keyboard.h"