+++ arch/x86_32/core/mouse.c
@@ -19,7 +19,7 @@ along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
 #include <stdbool.h>
 
-#include "arch/stubs.h"
+#include "hal.h"
 
 #include "drivers/mouse.h"