+++ arch/x86_32/core/pci.c
@@ -19,7 +19,7 @@ along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
 #include <stdint.h>
 
-#include "arch/stubs.h"
+#include "hal.h"
 
 #include "cpu/pci.h"