+++ kernel/drivers/storage/ata.c
@@ -19,7 +19,8 @@ along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
 #include <stdint.h>
 
-#include "arch/stubs.h"
+#include "hal.h"
+
 #include "cpu/pci.h"
 #include "drivers/storage/bdl.h"
 #include "drivers/terminal.h"